Struct finchers_template::Renderer

pub struct Renderer<T> { /* fields omitted */ }

The type representing a renderer using the specified template engine.

Methods

impl<T> Renderer<T> where
T: TemplateEngine,

pub fn content_type(self, content_type: Mime) -> Renderer<T>

Set the content-type of generated content.

By default, the value is guessed from the name of template.